requirements.txt
flask
transformers
torch
pydantic
flask
accelerate
sentencepiece
protobuf
pytest
python-dotenv